+++ R/get_c14data.R
@@ -132,6 +132,7 @@ get_all_parser_functions <- function() {
     "austarch" =  c14bazAAR::get_austarch,
     "calpal" = c14bazAAR::get_calpal,
     "context" = c14bazAAR::get_context,
+    "emedyd" = c14bazAAR::get_emedyd,
     "eubar" = c14bazAAR::get_eubar,
     "euroevol" = c14bazAAR::get_euroevol,
     "radon" = c14bazAAR::get_radon,
@@ -143,7 +144,8 @@ get_all_parser_functions <- function() {
     "14cpalaeolithic" = c14bazAAR::get_14cpalaeolithic,
     "medafricarbon" = c14bazAAR::get_medafricarbon,
     "jomon" = c14bazAAR::get_jomon,
-    "mesorad" = c14bazAAR::get_mesorad
+    "mesorad" = c14bazAAR::get_mesorad,
+    "katsianis" = c14bazAAR::get_katsianis
   ))
 }