+++ NEWS.md
@@ -1,10 +1,26 @@
-# c14bazAAR 1.2.1
+# c14bazAAR 1.3.0
+
+## general changes
+- added a CITATION file (see citation("c14bazAAR"))
+- deprecated `mark_duplicates` to get rid of this extra step. You can get the same result now with `remove_duplicates(mark_only = TRUE)` (#100)
+- deprecated `coordinate_precision`. The functionality was not essential and the calculated precision values probably frequently misleading. Beyond that the name was confusing (#96) (#106)
+- deprecated `finalize_country_name`. This wrapper function was rather confusing and the functionality can be very easily be reimplemented if necessary (#96) (#106)
+- renamed `standardize_country_name` to `fix_database_country_name` to make more clear what it does (#96) (#106)
 
 ## new getter functions
-- added getter function for PACEA database: `get_pacea`
-- added getter function for 14C-Palaeolithic database: `get_14cpalaeolithic`
-- added getter function for MedAfriCarbon database: `get_medafricarbon`
-- added getter function for Jōmon population dynamics database: `get_jomon`
+- added getter function for Katsianis et al. database: `get_katsianis` (#103)
+- added getter function for PACEA database: `get_pacea` (#90)
+- added getter function for 14C-Palaeolithic database: `get_14cpalaeolithic` (#90)
+- added getter function for MedAfriCarbon database: `get_medafricarbon` (#95)
+- added getter function for Jōmon population dynamics database: `get_jomon` (#95)
+- added getter function for emedyd database: `get_emedyd` (#102)
+
+## database updates
+- updated the CalPal database from version 2017_04_27 to 2020_08_20 (#108)
+
+## bugfixes
+- `lwgeom::st_make_valid` was replaced by `sf::st_make_valid` (#99)
+- enabled UTF-8 characters in country thesaurus (#96) (#104)
 
 # c14bazAAR 1.2.0
 
@@ -16,7 +32,6 @@
 - added a basic plot function for c14_date_lists (#82)
 - added a basic write function for c14_date_lists: `write_c14()` (#84)
 - added a version column that documents from which database version a certain date is pulled (#85)
-
 
 # c14bazAAR 1.1.0
 
@@ -47,7 +62,6 @@
 ## new getter functions
 - added getter function for Palmisano et al. database: `get_palmisano` (#59)
 - added getter function for eubar database: `get_eubar` (#64)
-
 
 ## new features
 - added new options for the deduplication function (see `?duplicates`) (#63)